+++ src/commands/list.rs
@@ -1,10 +1,12 @@
 use std::collections::HashMap;
-use std::path::Path;
+use std::fs::File;
+use std::io::{stdout, Write};
+use std::path::{Path, PathBuf};
 use crate::aliases::{ModFileName, ModID};
 use crate::api::api_structs::{ModInfo};
 use crate::commands::sync::{get_sync_data, sync, ModSyncInfo};
 use crate::rustique_errors::RustiqueError;
-use crate::utils::{extract_all_mods_metadata, gather_dependencies, gather_missing_dependencies, split_modid_version, sanitize_string};
+use crate::utils::{extract_all_mods_metadata, gather_dependencies, gather_missing_dependencies, split_modid_version, sanitize_string, format_for_csv, html_parse};
 use crate::version_management::parse_version;
 use owo_colors::OwoColorize;
 use comfy_table::modifiers::UTF8_ROUND_CORNERS;
@@ -12,12 +14,17 @@ use comfy_table::presets::UTF8_FULL;
 use comfy_table::{Cell, CellAlignment, ContentArrangement, Row, Table};
 use std::str::FromStr;
 use std::time::Instant;
+use csv::Writer;
 use tracing::{debug, info};
+use crate::commands::arg_structs::list_args::ListOutput;
+use crate::commands::info::info;
+use crate::commands::search::parse_search_file;
 use crate::config::config_manager::get_config;
-use crate::config::config_structs::{CellAttr, CellColor, ListColumn};
+use crate::config::config_structs::{CellAttr, CellColor, ListColumn, TableSection};
 use crate::information_utils::prep_cell;
 use crate::install_manager::Install;
 use crate::traits::ref_ext::PathRef;
+use crate::traits::string_ext::StrLowerExt;
 
 fn grab_this_mod_deps(mod_info: &ModInfo, dep_list: &[Install]) -> String {
     let mut res = dep_list.iter()
@@ -29,18 +36,37 @@ fn grab_this_mod_deps(mod_info: &ModInfo, dep_list: &[Install]) -> String {
 }
 
 #[allow(clippy::filter_map_next, clippy::too_many_lines)]
-pub async fn cmd_list(mod_dir: impl PathRef, only_updated: bool, modpack_call: bool, local_mp_call: bool) -> Result<(), RustiqueError> {
+pub async fn cmd_list(mod_dir: impl PathRef, only_updated: bool, modpack_call: bool, local_mp_call: bool, columns: Vec<ListColumn>, output: Option<ListOutput>, write_file: Option<PathBuf>) -> Result<(), RustiqueError> {
     let mod_dir = mod_dir.as_ref();
     let start_time = Instant::now();
     let config = get_config().read().await;
 
-    let list_columns = &config.table.list;
+    let config_columns = &config.table.list;
+
 
     // setup headers
     let mut table = Table::new();
     table.load_preset(UTF8_FULL)
         .apply_modifier(UTF8_ROUND_CORNERS)
         .set_content_arrangement(ContentArrangement::Dynamic);
+
+    // if fields set the columns from that
+    // else get columns from config
+
+    let mut override_columns = TableSection::new();
+
+    let list_columns = if columns.is_empty() {
+        config_columns
+    } else {
+
+        for col in columns {
+            override_columns.headers.with(col.as_str(), None, None);
+            override_columns.cells.with(col.as_str(), None,None);
+        }
+
+        &override_columns
+    };
+
 
     let header_cells : Vec<Cell> = list_columns.headers.iter().filter_map(|(column, properties)| {
         let color = properties.color.clone();
@@ -62,16 +88,18 @@ pub async fn cmd_list(mod_dir: impl PathRef, only_updated: bool, modpack_call: b
             Ok(ListColumn::LastUpdateLocal) => "Last Update (Local)",
             Ok(ListColumn::LastUpdateRemote)    => "Last Update (Remote)",
             Ok(ListColumn::Website)             => "Website",
+            Ok(ListColumn::ModURL)              => "Mod URL",
             _ => "N/A"
         };
-        
+
+
         if (local_mp_call || modpack_call) && matches!(ListColumn::from_str(column), Ok(ListColumn::MissingDeps)) {
             debug!("modpack_call on missingDeps");
             return None;
         }
         
         if local_mp_call && matches!(ListColumn::from_str(column), Ok(ListColumn::LatestVersion)) {
-            debug!("local modpack_call on LatestVersio");
+            debug!("local modpack_call on LatestVersion");
             return None; 
         }
         
@@ -107,21 +135,34 @@ pub async fn cmd_list(mod_dir: impl PathRef, only_updated: bool, modpack_call: b
     };
     
     let missing_deps = gather_missing_dependencies(&installed_mods, &[], sync_hashmap);
-   
-    
+
     let mut enabled_modpacks: HashMap<ModID, Vec<ModID>> = config.modpacks.enabled.iter().map(|m| (m.clone(), Vec::new())).collect();
+
+    let search_db = parse_search_file().await?;
     
     for (pack_id, v) in &mut enabled_modpacks {
         let (pack_id, _) = split_modid_version(pack_id);
         let mpath = Path::new(&config.modpacks.modpack_dir).join("installed").join(pack_id);
         if mpath.exists() {
-            let mp_sync_file = if let Ok(s) = get_sync_data(&mpath, true).await {
-                info!("Sync data found!");
-                s
-            } else {
-                info!("Sync data bad, repopulating");
-                sync(mod_dir, true, vec![]).await?
+            // let mp_sync_file = if let Ok(s) = get_sync_data(&mpath, true).await {
+            //     info!("Sync data found!");
+            //     s
+            // } else {
+            //     info!("Sync data bad, repopulating");
+            //     sync(mod_dir, true, vec![]).await?
+            // };
+            
+            let mp_sync_file = match get_sync_data(&mpath, false).await {
+                Ok(s) => {
+                    info!("Sync data found");
+                    s
+                },
+                Err(e) => { 
+                    info!("{}: {}", "ERROR: ".red().bold(), e);
+                    sync(mod_dir, true, vec![]).await? 
+                }
             };
+            
             let keys: Vec<ModID> = mp_sync_file.rustique_sync.into_keys().map(|k| split_modid_version(k).0).collect();
             v.extend(keys);
         }
@@ -144,12 +185,13 @@ pub async fn cmd_list(mod_dir: impl PathRef, only_updated: bool, modpack_call: b
                 .is_some_and(|sync| sync.latest_known_version != sync.installed_version)
         })
         .map(|(filename, mod_info)| {
-            
-            
-            let file_is_smylink = mod_dir.join(filename).is_symlink();
+
+
+
+            let file_is_symlink = mod_dir.join(filename).is_symlink();
            
             let pkg = config.pkg.iter().find(|p| p.mod_id.eq(&mod_info.mod_id));
-            let cells: Vec<Cell> = list_columns.cells.iter().filter_map(|(column, properties)| { 
+            let cells: Vec<Cell> = list_columns.cells.iter().filter_map(|(column, properties)| {
                 
                 let color = properties.color.clone();
                 let attr = properties.attribute.clone();
@@ -165,6 +207,15 @@ pub async fn cmd_list(mod_dir: impl PathRef, only_updated: bool, modpack_call: b
                         None
                     }
                 }).next().unwrap_or_default();
+
+
+                let mod_api_search = search_db.mods.iter().find(|m| {
+                    if let Some(name) = &m.name {
+                        name.contains_str_only(&mod_sync_data.mod_name) || name.contains_str_only(mod_sync_id.as_str())
+                    } else {
+                        false
+                    }
+                });
 
                 match <ListColumn as FromStr>::from_str(column) {
                     Ok(ListColumn::Name) => {
@@ -187,7 +238,7 @@ pub async fn cmd_list(mod_dir: impl PathRef, only_updated: bool, modpack_call: b
                             the_color = Some(CellColor::Green);
                         }
                         
-                        if file_is_smylink {
+                        if file_is_symlink {
                             txt += "(";
                             txt += enabled_modpacks.iter()
                                 .find(|(_,v)| v.contains(&mid))
@@ -230,20 +281,43 @@ pub async fn cmd_list(mod_dir: impl PathRef, only_updated: bool, modpack_call: b
                             ))
                     }
                     Ok(ListColumn::Description) => {
-                        let txt = sanitize_string(&mod_info.description.clone().unwrap_or(String::new()));
+                        let mut txt = sanitize_string(&mod_info.description.clone().unwrap_or(String::new()));
+
+                        if let Some(out) = &output {
+                            if matches!(out, ListOutput::Csv) {
+                                txt = format_for_csv(txt);
+                            }
+                        }
+
                         Some(prep_cell(&txt, color, attr, None, None))
                     },
                     Ok(ListColumn::Deps) => {
-                        let deps = grab_this_mod_deps(mod_info, &all_deps.clone());
-                        Some(prep_cell(&deps, color, attr, Some(','), None))
+                        let mut deps = grab_this_mod_deps(mod_info, &all_deps.clone());
+                        let mut table_sep = Some(',');
+                        if let Some(out) = &output {
+                           if matches!(out, ListOutput::Csv) {
+                               deps = format_for_csv(deps);
+                               table_sep = None;
+                           }
+                        }
+
+                        Some(prep_cell(&deps, color, attr, table_sep, None))
                     }
                     Ok(ListColumn::MissingDeps) => {
                         if modpack_call {
                             return None;
                         }
-                        let missing = grab_this_mod_deps(mod_info, &missing_deps.clone());
+                        let mut missing = grab_this_mod_deps(mod_info, &missing_deps.clone());
+                        let mut table_sep = Some(',');
+                        // if output === csv, replace all , with space
+                        if let Some(out) = &output {
+                           if matches!(out, ListOutput::Csv) {
+                               missing = format_for_csv(missing);
+                               table_sep = None;
+                           }
+                        }
                         
-                        Some(prep_cell(&missing, color, attr, Some(','), None))
+                        Some(prep_cell(&missing, color, attr, table_sep, None))
                     }
                     Ok(ListColumn::Filename) => {
                         Some(prep_cell(filename.as_str(), color, attr, None, None))
@@ -254,6 +328,12 @@ pub async fn cmd_list(mod_dir: impl PathRef, only_updated: bool, modpack_call: b
                         let gv  = mod_sync_data.game_versions;
                         let game_versions = if gv.len() > 1  {
                             format!("{} - {}", gv[gv.len() - 1], gv[0])
+                        } else if let Some(out) = &output {
+                            if matches!(out, ListOutput::Csv) {
+                                gv.join(" ")
+                            } else {
+                                gv.join(",")
+                            }
                         } else {
                             gv.join(",")
                         };
@@ -261,13 +341,40 @@ pub async fn cmd_list(mod_dir: impl PathRef, only_updated: bool, modpack_call: b
                     }
                     Ok(ListColumn::LastUpdateLocal 
                        | ListColumn::LastUpdateRemote 
-                       | ListColumn::HasBackup 
-                       | ListColumn::Changelog) => {
+                       | ListColumn::HasBackup) => {
                         Some(prep_cell("NOT IMPLEMENTED", color, attr, None, None))
+                    }
+                    Ok(ListColumn::Changelog) => {
+                        // let mut changelog = sanitize_string(&mod_sync_data.latest_changelog);
+                        let mut changelog = &mod_sync_data.latest_changelog;
+
+                        let out =  html_parse(&mut changelog, usize::from(table.width().unwrap_or_default())).unwrap_or_default();
+                        let changelog = if let Some(output) = &output {
+                            if matches!(output, ListOutput::Csv) {
+                                format_for_csv(out)
+                            } else {
+                                out
+                            }
+                        } else {
+                           out
+                        };
+                        
+                        Some(prep_cell(changelog, color, attr, None, None))
                     }
                     Ok(ListColumn::Website) => {
                         Some(prep_cell(mod_info.website.clone().unwrap_or_default().as_str(), color, attr, None, None))
                     },
+                    Ok(ListColumn::ModURL) => {
+
+
+                        let url = if let Some(api) = mod_api_search {
+                            format!("https://mods.vintagestory.at/show/mod/{}", api.asset_id)
+                        } else {
+                            "Can't determine".into()
+                        };
+
+                        Some(prep_cell(url, color, attr, None, None))
+                    }
                     _ => Some(prep_cell("", color, attr, None, None))
                 } 
             }).collect();
@@ -276,15 +383,40 @@ pub async fn cmd_list(mod_dir: impl PathRef, only_updated: bool, modpack_call: b
     }).collect();
 
     table.add_rows(rows);
+    
+    if let Some(out) = &output {
+        if matches!(out, ListOutput::Csv) {
+            
+           
+            let writer: Box<dyn Write> = match write_file {
+                Some(f) => Box::new(File::create(f)?),
+                None => Box::new(stdout()),
+            };
+            
+            let mut wtr = Writer::from_writer(writer);
+            
+            if let Some(row) = table.header() {
+                let h: Vec<String> = row.cell_iter().map(|c|c.content()).collect();
+                wtr.write_record(h).map_err(|e| RustiqueError::SimpleError(e.to_string()))?
+            }
 
-    println!("{table}");
-    print!("{} {}", "Total Mod Count:".bright_green().bold().on_black(), installed_mods.len().to_string().bright_purple().on_black());
-
-     if config.show_execution_time {
-        let elapsed = format!("{:.2}", start_time.elapsed().as_secs_f64());
-        println!(" - {}: {}{}","List operation took".bright_green().bold().on_black(), elapsed.bright_purple().on_black(), "s".bright_yellow().on_black());
+            for row in table.row_iter() {
+                let r: Vec<String> = row.cell_iter().map(|c|c.content()).collect();
+                wtr.write_record(r).map_err(|e| RustiqueError::SimpleError(e.to_string()))?
+            }
+            
+           
+            wtr.flush()?;
+        }
     } else {
-        println!();
+
+        println!("{table}");
+        print!("{} {}", "Total Mod Count:".bright_green().bold().on_black(), installed_mods.len().to_string().bright_purple().on_black());
+
+        if config.show_execution_time {
+            let elapsed = format!("{:.2}", start_time.elapsed().as_secs_f64());
+            println!(" - {}: {}{}","List operation took".bright_green().bold().on_black(), elapsed.bright_purple().on_black(), "s".bright_yellow().on_black());
+        }
     }
 
     Ok(())