+++ src/install_manager.rs
@@ -215,7 +215,7 @@ pub async fn install_manager(
                 mod_to_install.mod_name = res_mod.mod_json.name.clone().unwrap_or_default();
                 
                 let pkg = config.pkg.iter().find(|p| p.mod_id.eq(&res_mod.mod_json.mod_id.to_string()));
-                let (mod_version, download_url, _) = if let Some(mod_pkg) = pkg {
+                let (mod_version, download_url, _,_) = if let Some(mod_pkg) = pkg {
                     parse_pinned_version(&res_mod.mod_json.releases, &mod_pkg.clone(), config.pinned_game_version.clone())
                 } else {
                     parse_latest_version(&res_mod.mod_json.releases)